Fixup DATABASE_URL
The connection string used by the ragtime migration configuration in the
project file differs from the connection string provided in the
DATABASE_URL string on Heroku.  Without this patch the application is
unable to run:

    heroku run 'lein do clean; lein ragtime migrate;'

This patch addresses the problem by parsing the DATABASE_URL and
transforming it into a connection string ragtime will accept from the
project.clj configuration.